Fix NPE bug in materialize rewrite
